Use a RegEx to get carrier instead of XML parsing
As we later do literal matches against the lines XML parsing is the
wrong approach as e.g. XML excapes will not be found which causes custom
APNs with e.g. `&amp;` to be ignored, i.e. they will not overwrite the
default ones and will be fully missing from the output file.

Change-Id: I2bc575d4bbdc5d802c5d5b3420ee6b536a5a18fc